+++ packages/core/src/index.ts
@@ -19,7 +19,7 @@ export { Logger, LoggerConfig, LoggerMethod } from './types/logger';
 export { ControllerErrorHandler } from './services/controller-error-handler';
 export { BodyParserConfig } from './models/body-parser-config';
 export { ModConfig } from './models/mod-config';
-export { DefaultLogger } from './services/default-logger';
+export { ConsoleLogger } from './services/console-logger';
 export { Router } from './types/router';
 export { Status, getStatusText, isSuccess, STATUS_CODE_INFO } from './utils/http-status-codes';
 export { NodeResponse, NodeRequest, RequestListener } from './types/server-options';